refactor(core): split content persistence services
Move scoped client access, models, lineage, raw persistence, recall, and
document search into acyclic owners behind the stable content contract.

Preserve public signatures, SurrealQL, revision and lifecycle behavior,
organization scoping, embedding fallback, and client concurrency semantics.
