+++ apps/api/tests/conftest.py
@@ -10,7 +10,7 @@ import pytest
 @pytest.fixture(autouse=True)
 def disable_raw_memory_auto_embedding(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(
-        "sibyl_core.services.surreal_content._configured_raw_memory_embedding_provider",
+        "sibyl_core.services.content_models.configured_raw_memory_embedding_provider",
         lambda: None,
     )
