+++ packages/python/sibyl-core/tests/conftest.py
@@ -44,7 +44,7 @@ from sibyl_core.models.tasks import (
 @pytest.fixture(autouse=True)
 def disable_raw_memory_auto_embedding(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(
-        "sibyl_core.services.surreal_content._configured_raw_memory_embedding_provider",
+        "sibyl_core.services.content_models.configured_raw_memory_embedding_provider",
         lambda: None,
     )
